configuração da controller para CRUD de usuarios

+++ src/main/java/sptech/school/repositories/UserRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import sptech.school.entities.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     User findByEmail(String email);
 }
